added sorting and extracted peers functionality
